requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
google-auth==2.16.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
oracledb==2.1.1
Faker==18.9.0
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0